Moved tests to SuperAdminServiceImplTest class
